[frontend] dashboard widget refacto
