+++ opencti-platform/opencti-front/lang/front/de.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "Diese Regel wurde auf die vorhandenen Daten angewandt",
   "This score is updated with the decay rule applied to this indicator.": "Diese Punktzahl wird mit der auf diesen Indikator angewandten Verfallsregel aktualisiert.",
   "This sector does not have any description.": "Für diesen Sektor gibt es keine Beschreibung.",
+  "This step is not implemented": "Dieser Schritt ist nicht implementiert",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "Dieser Trigger/Digest wurde für Sie freigegeben und Sie können ihn nicht ändern oder löschen",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "Dieser Benutzer erbt keine maximale Vertrauensstufe von seiner Gruppe. Konfigurieren Sie Benutzer's Gruppen mit einem maximalen Vertrauensniveau.",
   "This user has no effective confidence level from the groups assigned.": "Dieser Benutzer hat kein effektives Vertrauensniveau der zugewiesenen Gruppen.",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "Dieser Benutzer hat kein Max Confidence Level und erbt auch keines von Gruppen. Fügen Sie eine Gruppe zu diesem Benutzer hinzu, um das Problem zu beheben.",
+  "This value cannot be null": "Dieser Wert kann nicht null sein",
   "This value is overridden for the following entity types": "Dieser Wert wird für die folgenden Entitätstypen außer Kraft gesetzt",
+  "This widget type is not implemented": "Dieser Widget-Typ ist nicht implementiert",
   "Threat actor": "Bedrohungsakteur",
   "Threat actor types": "Arten von Bedrohungsakteuren",
   "Threat actors": "Bedrohungsakteure",
