+++ opencti-platform/opencti-front/lang/front/en.json
@@ -2856,11 +2856,14 @@
   "This rule has been applied on the existing data": "This rule has been applied on the existing data",
   "This score is updated with the decay rule applied to this indicator.": "This score is updated with the decay rule applied to this indicator.",
   "This sector does not have any description.": "This sector does not have any description.",
+  "This step is not implemented": "This step is not implemented",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "This trigger/digest has been shared with you and you are not able to modify or delete it",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.",
   "This user has no effective confidence level from the groups assigned.": "This user has no effective confidence level from the groups assigned.",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.",
+  "This value cannot be null": "This value cannot be null",
   "This value is overridden for the following entity types": "This value is overridden for the following entity types",
+  "This widget type is not implemented": "This widget type is not implemented",
   "Threat actor": "Threat actor",
   "Threat actor types": "Threat actor types",
   "Threat actors": "Threat actors",
