+++ opencti-platform/opencti-front/lang/front/es.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "Esta regla se ha aplicado a los datos existentes",
   "This score is updated with the decay rule applied to this indicator.": "Esta puntuación se actualiza con la regla de decaimiento aplicada a este indicador.",
   "This sector does not have any description.": "Este sector no tiene ninguna descripción.",
+  "This step is not implemented": "Este paso no está implementado",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "Este desencadenante/digerido ha sido compartido con usted y no puede modificarlo ni eliminarlo",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "Este usuario no hereda un nivel de confianza máximo de su grupo. Configurar grupos de usuarios con un nivel de confianza máximo.",
   "This user has no effective confidence level from the groups assigned.": "Este usuario no tiene nivel de confianza efectivo de los grupos asignados.",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "Este usuario no tiene Nivel de Confianza Máximo y no hereda uno de los grupos. Agregue un grupo a este usuario para resolver el problema.",
+  "This value cannot be null": "Este valor no puede ser nulo",
   "This value is overridden for the following entity types": "Este valor se anula para los siguientes tipos de entidad",
+  "This widget type is not implemented": "Este tipo de widget no está implementado",
   "Threat actor": "Actor de amenazas",
   "Threat actor types": "Tipos de actor de amenazas",
   "Threat actors": "Actores de amenazas",
