+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "Cette règle a été appliquée sur les données de la plateforme",
   "This score is updated with the decay rule applied to this indicator.": "Ce score est mis à jour avec la règle de désintégration appliquée à cet indicateur.",
   "This sector does not have any description.": "Ce secteur n'a aucune description.",
+  "This step is not implemented": "Cette étape n'est pas mise en œuvre",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "Ce déclencheur/digest a été partagé avec vous et vous n'êtes pas en mesure de le modifier ou de le supprimer",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "Cet utilisateur n'hérite pas d'un niveau de confiance maximum de son groupe. Configurer les groupes d'utilisateurs avec un niveau de confiance maximum.",
   "This user has no effective confidence level from the groups assigned.": "Cet utilisateur n'a pas de niveau de confiance effectif dans les groupes qui lui ont été attribués.",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "Cet utilisateur n'a pas de niveau de confiance maximal et n'en hérite pas des groupes. Ajoutez un groupe à cet utilisateur pour résoudre le problème.",
+  "This value cannot be null": "Cette valeur ne peut être nulle",
   "This value is overridden for the following entity types": "Cette valeur est remplacée pour les types d'entités suivants",
+  "This widget type is not implemented": "Ce type de widget n'est pas implémenté",
   "Threat actor": "Acteur",
   "Threat actor types": "Type d'acteur",
   "Threat actors": "Acteurs",
