+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "既存のデータにルールを適用しました",
   "This score is updated with the decay rule applied to this indicator.": "このスコアはこのインジケータに適用される減衰ルールで更新される。",
   "This sector does not have any description.": "このセクターには説明がありません。",
+  "This step is not implemented": "このステップは実装されていません",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "このトリガー/ダイジェストはあなたと共有され、あなたはそれを変更または削除することはできません。",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "このユーザはグループから最大信頼レベルを継承しません。最大信頼レベルを持つユーザーグループを設定します。",
   "This user has no effective confidence level from the groups assigned.": "このユーザーには、割り当てられたグループからの有効な信頼レベルがありません。",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "このユーザーには最大信頼度がなく、グループからの信頼度も継承していません。この問題を解決するには、このユーザーにグループを追加してください。",
+  "This value cannot be null": "この値をNULLにすることはできない",
   "This value is overridden for the following entity types": "この値は以下のエンティティ・タイプでオーバーライドされます。",
+  "This widget type is not implemented": "このウィジェットタイプは実装されていない",
   "Threat actor": "脅威アクター",
   "Threat actor types": "脅威アクターの種別",
   "Threat actors": "脅威行為者",
