+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "이 규칙은 기존 데이터에 적용되었습니다",
   "This score is updated with the decay rule applied to this indicator.": "이 점수는 이 인디케이터에 적용된 감소 규칙으로 업데이트됩니다.",
   "This sector does not have any description.": "이 부문에는 설명이 없습니다.",
+  "This step is not implemented": "이 단계는 구현되지 않습니다",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "이 트리거/다이제스트가 귀하와 공유되었으며 수정 또는 삭제할 수 없습니다",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "이 사용자는 그룹에서 최대 신뢰 수준을 상속받지 않습니다. 사용자 그룹을 최대 신뢰 수준으로 구성하십시오.",
   "This user has no effective confidence level from the groups assigned.": "이 사용자는 할당된 그룹에서 유효한 신뢰 수준이 없습니다.",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "이 사용자는 최대 신뢰 수준이 없으며 그룹에서 상속받지 않습니다. 이 사용자를 그룹에 추가하여 문제를 해결하십시오.",
+  "This value cannot be null": "이 값은 null일 수 없습니다",
   "This value is overridden for the following entity types": "이 값은 다음 엔터티 유형에 대해 재정의됩니다",
+  "This widget type is not implemented": "이 위젯 유형은 구현되지 않았습니다",
   "Threat actor": "위협 행위자",
   "Threat actor types": "위협 행위자 유형",
   "Threat actors": "위협 행위자",
