+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -2852,11 +2852,14 @@
   "This rule has been applied on the existing data": "该规则已应用于现有数据",
   "This score is updated with the decay rule applied to this indicator.": "该分数会根据适用于该指标的衰减规则进行更新。",
   "This sector does not have any description.": "此部门没有任何描述。",
+  "This step is not implemented": "此步骤未执行",
   "This trigger/digest has been shared with you and you are not able to modify or delete it": "这个触发器/摘要已经与你共享，你不能修改或删除它。",
   "This user does not inherit a Max Confidence Level from their group. Configure user's groups with a Max Confidence Level.": "该用户不继承其组的最大置信度。为用户组配置最大置信度。",
   "This user has no effective confidence level from the groups assigned.": "该用户没有指定组的有效信任级别。",
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "该用户没有最大信任度，也没有从组中继承最大信任度。为该用户添加一个组来解决问题。",
+  "This value cannot be null": "该值不能为空",
   "This value is overridden for the following entity types": "以下实体类型可重载该值",
+  "This widget type is not implemented": "未执行此部件类型",
   "Threat actor": "威胁主体",
   "Threat actor types": "威胁主体类型",
   "Threat actors": "威胁行为者",
