+++ opencti-platform/opencti-front/src/private/components/widgets/widgetExportHandler.tsx
@@ -1,7 +1,7 @@
 import fileDownload from 'js-file-download';
 import { graphql } from 'react-relay';
-import { widgetExportHandlerQuery$data } from '@components/workspaces/dashboards/__generated__/widgetExportHandlerQuery.graphql';
-import { fetchQuery } from '../../../../relay/environment';
+import type { widgetExportHandlerQuery$data } from './__generated__/widgetExportHandlerQuery.graphql';
+import { fetchQuery } from '../../../relay/environment';
 
 interface widgetToExport {
   id: string;
