+++ opencti-platform/opencti-front/src/private/components/workspaces/dashboards/Dashboard.jsx
@@ -12,10 +12,10 @@ import WorkspaceHeader from '../WorkspaceHeader';
 import { commitMutation } from '../../../../relay/environment';
 import { workspaceMutationFieldPatch } from '../WorkspaceEditionOverview';
 import useGranted, { EXPLORE_EXUPDATE } from '../../../../utils/hooks/useGranted';
-import WidgetPopover from './WidgetPopover';
+import WorkspaceWidgetPopover from './WorkspaceWidgetPopover';
 import { fromB64, toB64 } from '../../../../utils/String';
-import WidgetConfig from './WidgetConfig';
-import WidgetText from './WidgetText';
+import WorkspaceWidgetConfig from './WorkspaceWidgetConfig';
+import WidgetText from '../../widgets/WidgetText';
 import StixCoreObjectsMultiVerticalBars from '../../common/stix_core_objects/StixCoreObjectsMultiVerticalBars';
 import StixCoreObjectsNumber from '../../common/stix_core_objects/StixCoreObjectsNumber';
 import StixCoreObjectsList from '../../common/stix_core_objects/StixCoreObjectsList';
@@ -841,7 +841,7 @@ const DashboardComponent = ({ workspace, noToolbar }) => {
               variant="outlined"
                    >
               {!noToolbar && (
-              <WidgetPopover
+              <WorkspaceWidgetPopover
                 widget={widget}
                 manifest={manifest}
                 workspace={workspace}
@@ -909,7 +909,7 @@ const DashboardComponent = ({ workspace, noToolbar }) => {
           })}
         </ReactGridLayout>
       )}
-      {!noToolbar && userCanEdit && !isFeatureEnable('FAB_REPLACEMENT') && <WidgetConfig onComplete={handleAddWidget} workspace={workspace} />}
+      {!noToolbar && userCanEdit && !isFeatureEnable('FAB_REPLACEMENT') && <WorkspaceWidgetConfig onComplete={handleAddWidget} workspace={workspace} />}
     </div>
   );
 };
